backend/requirements.txt
FastAPI==0.124.0
uvicorn
pydantic==2.12.5
python-dotenv==1.2.1
faiss-cpu==1.13.1
sentence-transformers
open-clip-torch==2.27.0
Pillow>=10.0.0
numpy
redis==7.1.0
psycopg2-binary==2.9.11
SQLAlchemy
pandas
torch
scipy
scikit-learn
python-jose
passlib[bcrypt]
PyJWT
httpx
requests
safetensors==0.7.0
scikit-learn==1.7.2
scipy==1.16.3
sentence-transformers==5.1.2
setuptools==80.9.0
six==1.17.0
starlette==0.50.0
sympy==1.14.0
threadpoolctl==3.6.0
tokenizers==0.22.1
torch==2.9.1
tqdm==4.67.1
transformers==4.57.3
triton==3.5.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.6.1
uvicorn==0.38.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==15.0.1
